pyproject.toml
[tool.poetry]
name = "llm-sentence-ordering"
version = "0.1.0"
description = "Evaluating LLMs on sentence ordering"
packages = [
    { include = "dataset", from = "src" },
    { include = "eval", from = "src" },
]
authors = ["Shawon Ashraf"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
create-dataset = "dataset.pipelines:pipeline"
evaluate = "eval.evaluate:run"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.0.0"
wandb = "^0.17.2"
openai = "^1.34.0"
jupyter = "^1.0.0"
matplotlib = "^3.9.0"
polars = "^0.20.31"
pydantic = "^2.7.4"
datasets = "^2.20.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
jinja2 = "^3.1.4"
instructor = "^1.3.3"
scipy = "^1.14.0"
fire = "^0.6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
